NetBSD support with envsys ioctl and plist (#69)
* NetBSD support with envsys ioctl and plist

New examples, simple.rs only refreshed the first battery, and this does not
help checking for iterator validity for development, all.rs solves the
problem. The example oneshot.rs is just for comfort avoiding loop.

NetBSD exposes battery information through one ioctl so for now we have
to "waste" resource getting all sensors info for each single battery
refresh. We can't assume the library user will want to refresh all
bateries at the same time even if 90% of time it will probably be.

The result of the ioctl is a plist (apple format) that needs to be
parsed and accessed.

The plist library only uses options, so this is the reason for utils.rs.
These are all oneline functions to reduce the lenght of chains in the
netbsd platform device.rs file.

NetBSD does not expose battery vendor, model, technology, cycle among
other things for now, but these are optional.

Got inspiration from the FreeBSD port regarding having a better understanding on
how the library works.

Thanks to unitedbsd.com community for the guidance, and shout out to
https://github.com/distatus/battery/blob/master/battery_netbsd.go as
well as the NetBSD devs for the great man pages and the clear code
allowing to read what happens for undocumented things.

* Correcting rustfmt formatting on netbsd/device.rs

* Maybe crossbuild only as on FreeBSD ? (Already tried manually on Linux and it works)

* Revert "Maybe crossbuild only as on FreeBSD ? (Already tried manually on Linux and it works)"

Well the naive approach did not work

This reverts commit 693c88268575d18ee844659c0f5851ec5e37850e.

* Trying CICD regarding cross-rs

* Installing some prerequisites for NetBSD CI

* Correct path for libexecinfo.so in NetBSD CICD

* Clippy fixes for NetBSD platform

* Update src/platform/netbsd/iterator.rs

Remove default

Co-authored-by: David Knaack <davidkna@users.noreply.github.com>

* Update Cross.toml

Co-authored-by: David Knaack <davidkna@users.noreply.github.com>

* We support acpibat only in the end, but now we relyin on NetBSD kernel enum to ensure the data is correct even if someonechange envsys description fields. Extending the code for other battery driver is possible

* Updated the branch Cargo.toml to match updated version

---------

Co-authored-by: David Knaack <davidkna@users.noreply.github.com>